Move NameFormatterTab to gui.preftabs
